+++ Source/Data/Encoding/Gapotchenko.FX.Data.Encoding.Base16/Base16.cs
@@ -17,7 +17,7 @@ public sealed class Base16 : GenericBase16
     /// </summary>
     /// <param name="data">The byte span to encode.</param>
     /// <returns>The string with encoded data.</returns>
-    public new static string GetString(ReadOnlySpan<byte> data) => Instance.GetString(data);
+    public static new string GetString(ReadOnlySpan<byte> data) => Instance.GetString(data);
 
     /// <summary>
     /// Encodes all the bytes in the specified span into a string of Base16 characters with specified options.
@@ -25,7 +25,7 @@ public sealed class Base16 : GenericBase16
     /// <param name="data">The byte span to encode.</param>
     /// <param name="options">The options.</param>
     /// <returns>The string with encoded data.</returns>
-    public new static string GetString(ReadOnlySpan<byte> data, DataEncodingOptions options) => Instance.GetString(data, options);
+    public static new string GetString(ReadOnlySpan<byte> data, DataEncodingOptions options) => Instance.GetString(data, options);
 
     /// <summary>
     /// Decodes all Base16 characters in the specified read-only span into a byte array.
@@ -33,7 +33,7 @@ public sealed class Base16 : GenericBase16
     /// <param name="s">The read-only character span to decode.</param>
     /// <returns>A byte array with decoded data.</returns>
     /// <exception cref="FormatException">Cannot decode the input string.</exception>
-    public new static byte[] GetBytes(ReadOnlySpan<char> s) => Instance.GetBytes(s);
+    public static new byte[] GetBytes(ReadOnlySpan<char> s) => Instance.GetBytes(s);
 
     /// <summary>
     /// Decodes all Base16 characters in the specified read-only span into a byte array with specified options.
@@ -42,7 +42,7 @@ public sealed class Base16 : GenericBase16
     /// <param name="options">The options.</param>
     /// <returns>A byte array with decoded data.</returns>
     /// <exception cref="FormatException">Cannot decode the input string.</exception>
-    public new static byte[] GetBytes(ReadOnlySpan<char> s, DataEncodingOptions options) => Instance.GetBytes(s, options);
+    public static new byte[] GetBytes(ReadOnlySpan<char> s, DataEncodingOptions options) => Instance.GetBytes(s, options);
 
     /// <summary>
     /// Decodes all Base16 characters in the specified string into a byte array.