+++ Source/Data/Encoding/Gapotchenko.FX.Data.Encoding.Base32/Base32Hex.cs
@@ -17,7 +17,7 @@ public sealed class Base32Hex : GenericBase32
     /// </summary>
     /// <param name="data">The byte span to encode.</param>
     /// <returns>The string with encoded data.</returns>
-    public new static string GetString(ReadOnlySpan<byte> data) => Instance.GetString(data);
+    public static new string GetString(ReadOnlySpan<byte> data) => Instance.GetString(data);
 
     /// <summary>
     /// Encodes all the bytes in the specified span into a string of base32-hex characters with specified options.
@@ -25,7 +25,7 @@ public sealed class Base32Hex : GenericBase32
     /// <param name="data">The byte span to encode.</param>
     /// <param name="options">The options.</param>
     /// <returns>The string with encoded data.</returns>
-    public new static string GetString(ReadOnlySpan<byte> data, DataEncodingOptions options) => Instance.GetString(data, options);
+    public static new string GetString(ReadOnlySpan<byte> data, DataEncodingOptions options) => Instance.GetString(data, options);
 
     /// <summary>
     /// Decodes all base32-hex characters in the specified read-only span into a byte array.
@@ -33,7 +33,7 @@ public sealed class Base32Hex : GenericBase32
     /// <param name="s">The read-only character span to decode.</param>
     /// <returns>A byte array with decoded data.</returns>
     /// <exception cref="FormatException">Cannot decode the input string.</exception>
-    public new static byte[] GetBytes(ReadOnlySpan<char> s) => Instance.GetBytes(s);
+    public static new byte[] GetBytes(ReadOnlySpan<char> s) => Instance.GetBytes(s);
 
     /// <summary>
     /// Decodes all base32-hex characters in the specified read-only span into a byte array with specified options.
@@ -42,7 +42,7 @@ public sealed class Base32Hex : GenericBase32
     /// <param name="options">The options.</param>
     /// <returns>A byte array with decoded data.</returns>
     /// <exception cref="FormatException">Cannot decode the input string.</exception>
-    public new static byte[] GetBytes(ReadOnlySpan<char> s, DataEncodingOptions options) => Instance.GetBytes(s, options);
+    public static new byte[] GetBytes(ReadOnlySpan<char> s, DataEncodingOptions options) => Instance.GetBytes(s, options);
 
     /// <summary>
     /// Decodes all base32-hex characters in the specified string into a byte array.
@@ -73,14 +73,14 @@ public sealed class Base32Hex : GenericBase32
     /// </summary>
     /// <param name="s">The read-only character span to pad.</param>
     /// <returns>The padded encoded string.</returns>
-    public new static string Pad(ReadOnlySpan<char> s) => Instance.Pad(s);
+    public static new string Pad(ReadOnlySpan<char> s) => Instance.Pad(s);
 
     /// <summary>
     /// Unpads the encoded read-only character span.
     /// </summary>
     /// <param name="s">The read-only character span to unpad.</param>
     /// <returns>The unpadded read-only character span.</returns>
-    public new static ReadOnlySpan<char> Unpad(ReadOnlySpan<char> s) => Instance.Unpad(s);
+    public static new ReadOnlySpan<char> Unpad(ReadOnlySpan<char> s) => Instance.Unpad(s);
 
     [DebuggerBrowsable(DebuggerBrowsableState.Never)]
     static volatile IBase32? m_Instance;