+++ Source/Security/Cryptography/Gapotchenko.FX.Security.Cryptography/Arc4.cs
@@ -89,5 +89,5 @@ public abstract class Arc4 : SymmetricAlgorithm
     /// Creates an instance of the default implementation of ARC4 algorithm.
     /// </summary>
     /// <returns>The instance of ARC4 algorithm.</returns>
-    public new static Arc4 Create() => new Arc4Managed();
+    public static new Arc4 Create() => new Arc4Managed();
 }